rename docs
